Error errors from all endpoints more gracefully
