+++ assets/js/base/utils/errors.js
@@ -1,4 +1,9 @@
 /**
+ * External dependencies
+ */
+import { __ } from '@wordpress/i18n';
+
+/**
  * Given a JS error or a fetch response error, parse and format it so it can be displayed to the user.
  *
  * @param  {Object}   error           Error object.
@@ -27,4 +32,28 @@ export const formatError = async ( error ) => {
 		message: error.message,
 		type: error.type || 'general',
 	};
+};
+
+/**
+ * Given an API response object, formats the error message into something more human readable.
+ *
+ * @param  {Object}   response           Response object.
+ * @return {string}   Error message.
+ */
+export const formatStoreApiErrorMessage = ( response ) => {
+	if ( response.data && response.code === 'rest_invalid_param' ) {
+		const invalidParams = Object.values( response.data.params );
+		if ( invalidParams[ 0 ] ) {
+			return invalidParams[ 0 ];
+		}
+	}
+
+	if ( response?.message ) {
+		return response?.message;
+	}
+
+	return __(
+		'Something went wrong. Please contact us to get assistance.',
+		'woo-gutenberg-products-block'
+	);
 };
